cleanup / partial 'invert on activate'
The invert on activate bit is more annoying than it seems, but say
that the workspace background is mapped to a skydome or background
plane. Then set an external connection point in order to present on the
background, which is a nice way of watching videos. The problem then is,
as always, that the client origo might shift from upper-left to
lower-left, which might not match with the source image. This setup
would temporary-flip the background on connect, and then flip back when
lost.